MSG_PROBE_RESOURCE: change fd type from int32 to int64

+++ probe_badaapi/bada_file.cpp
@@ -188,7 +188,7 @@ result File::Construct(const Tizen::Base::String& filePath,
 		PACK_COMMON_BEGIN(MSG_PROBE_RESOURCE, API_ID_result_File__Construct_const_Tizen__Base__String__filePath__const_Tizen__Base__String__openMode__bool_createParentDirectories_,
 				  "ssd", temp_path, temp_mode, createParentDirectories);
 		PACK_COMMON_END(ret, ret, blockresult);
-		PACK_RESOURCE(0, (unsigned int)this, FD_FILE, FD_API_OPEN, 0, temp_path);
+		PACK_RESOURCE(0, (unsigned long)this, FD_FILE, FD_API_OPEN, 0, temp_path);
 		FLUSH_LOCAL_BUF();
 		 
 		 postBlockEnd();
@@ -389,7 +389,7 @@ result File::Construct(const Tizen::Base::String& filePath,
 				  API_ID_result_File__Construct_const_Tizen__Base__String__filePath__const_Tizen__Base__String__openMode_,
 				  "ss", temp_path, temp_mode);
 		PACK_COMMON_END(ret, ret, blockresult);
-		PACK_RESOURCE(0, (unsigned int)this, FD_FILE, FD_API_OPEN, 0, temp_path);
+		PACK_RESOURCE(0, (unsigned long)this, FD_FILE, FD_API_OPEN, 0, temp_path);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -514,7 +514,7 @@ result File::Construct(const Tizen::Base::String& filePath,
 				  API_ID_result_File__Construct_const_Tizen__Base__String__filePath__const_char__pOpenMode_,
 				  "sp", temp, pOpenMode);
 		PACK_COMMON_END(ret, ret, blockresult);
-		PACK_RESOURCE(0, (unsigned int)this, FD_FILE, FD_API_OPEN, 0, temp);
+		PACK_RESOURCE(0, (unsigned long)this, FD_FILE, FD_API_OPEN, 0, temp);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -567,7 +567,7 @@ result File::Construct(const Tizen::Base::String& filePath,
 				  API_ID_result_File__Construct_const_Tizen__Base__String__filePath__const_char__pOpenMode__const_Tizen__Base__ByteBuffer__secretKey_,
 				  "sp", temp, pOpenMode);
 		PACK_COMMON_END(ret, ret, blockresult);
-		PACK_RESOURCE(0, (unsigned int)this, FD_FILE, FD_API_OPEN, 0, temp);
+		PACK_RESOURCE(0, (unsigned long)this, FD_FILE, FD_API_OPEN, 0, temp);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -615,7 +615,7 @@ result File::Flush(void) {
 				  API_ID_result_File__Flush_void_,
 				  "", 0);
 		PACK_COMMON_END(ret, ret, blockresult);
-		PACK_RESOURCE(0, (unsigned int)this, FD_FILE, FD_API_OTHER, 0, temp);
+		PACK_RESOURCE(0, (unsigned long)this, FD_FILE, FD_API_OTHER, 0, temp);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -664,7 +664,7 @@ Tizen::Base::String File::GetName(void) const{
 				  API_ID_Tizen__Base__String_File__GetName_void__const,
 				  "", 0);
 		PACK_COMMON_END(ret.GetPointer(), res, blockresult);
-		PACK_RESOURCE(0, (unsigned int)this, FD_FILE, FD_API_OTHER, 0, temp);
+		PACK_RESOURCE(0, (unsigned long)this, FD_FILE, FD_API_OTHER, 0, temp);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -713,9 +713,9 @@ result File::Read(Tizen::Base::String& buffer) {
 		PREPARE_LOCAL_BUF();
 		PACK_COMMON_BEGIN(MSG_PROBE_RESOURCE,
 				  API_ID_result_File__Read_Tizen__Base__String__buffer_,
-				  "x", (unsigned int)&buffer);
+				  "x", (unsigned long)&buffer);
 		PACK_COMMON_END(ret, ret, blockresult);
-		PACK_RESOURCE(nRead, (unsigned int)this, FD_FILE, FD_API_READ, 0, temp);
+		PACK_RESOURCE(nRead, (unsigned long)this, FD_FILE, FD_API_READ, 0, temp);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -764,9 +764,9 @@ result File::Read(Tizen::Base::ByteBuffer& buffer) {
 		PREPARE_LOCAL_BUF();
 		PACK_COMMON_BEGIN(MSG_PROBE_RESOURCE,
 				  API_ID_result_File__Read_Tizen__Base__ByteBuffer__buffer_,
-				  "x", (unsigned int)&buffer);
+				  "x", (unsigned long)&buffer);
 		PACK_COMMON_END(ret, ret, blockresult);
-		PACK_RESOURCE(nRead, (unsigned int)this, FD_FILE, FD_API_READ, 0, temp);
+		PACK_RESOURCE(nRead, (unsigned long)this, FD_FILE, FD_API_READ, 0, temp);
 		FLUSH_LOCAL_BUF();
 		
 		 postBlockEnd();
@@ -816,9 +816,9 @@ int File::Read(void *buffer, int length) {
 		PREPARE_LOCAL_BUF();
 		PACK_COMMON_BEGIN(MSG_PROBE_RESOURCE,
 				  API_ID_int_File__Read_void__buffer__int_length_,
-				  "xd", (unsigned int)buffer, length);
+				  "xd", (unsigned long)buffer, length);
 		PACK_COMMON_END(ret, res, blockresult);
-		PACK_RESOURCE(nRead, (unsigned int)this, FD_FILE, FD_API_READ, 0, temp);
+		PACK_RESOURCE(nRead, (unsigned long)this, FD_FILE, FD_API_READ, 0, temp);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -879,7 +879,7 @@ result File::Seek(FileSeekPosition position, long offset) {
 				  API_ID_result_File__Seek_FileSeekPosition_position__long_offset_,
 				  "sx", temp_pos, offset);
 		PACK_COMMON_END(ret, ret, blockresult);
-		PACK_RESOURCE(0, (unsigned int)this, FD_FILE, FD_API_OTHER, 0, temp);
+		PACK_RESOURCE(0, (unsigned long)this, FD_FILE, FD_API_OTHER, 0, temp);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -929,7 +929,7 @@ int File::Tell(void) const {
 				  API_ID_int_File__Tell_void__const,
 				  "", 0);
 		PACK_COMMON_END(ret, res, blockresult);
-		PACK_RESOURCE(0, (unsigned int)this, FD_FILE, FD_API_OTHER, 0, temp);
+		PACK_RESOURCE(0, (unsigned long)this, FD_FILE, FD_API_OTHER, 0, temp);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -977,7 +977,7 @@ result File::Truncate(int length) {
 				  API_ID_result_File__Truncate_int_length_,
 				  "d", length);
 		PACK_COMMON_END(ret, ret, blockresult);
-		PACK_RESOURCE(0, (unsigned int)this, FD_FILE, FD_API_OTHER, 0, temp);
+		PACK_RESOURCE(0, (unsigned long)this, FD_FILE, FD_API_OTHER, 0, temp);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -1026,9 +1026,9 @@ result File::Write(const void *buffer, int length) {
 		PREPARE_LOCAL_BUF();
 		PACK_COMMON_BEGIN(MSG_PROBE_RESOURCE,
 				  API_ID_result_File__Write_const_void__buffer__int_length_,
-				  "xd", (unsigned int)buffer, length);
+				  "xd", (unsigned long)buffer, length);
 		PACK_COMMON_END(ret, ret, blockresult);
-		PACK_RESOURCE(nWritten, (unsigned int)this, FD_FILE, FD_API_WRITE, 0, temp);
+		PACK_RESOURCE(nWritten, (unsigned long)this, FD_FILE, FD_API_WRITE, 0, temp);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -1077,9 +1077,9 @@ result File::Write(const Tizen::Base::ByteBuffer& buffer) {
 		PREPARE_LOCAL_BUF();
 		PACK_COMMON_BEGIN(MSG_PROBE_RESOURCE,
 				  API_ID_result_File__Write_const_Tizen__Base__ByteBuffer__buffer_,
-				  "x", (unsigned int)&buffer);
+				  "x", (unsigned long)&buffer);
 		PACK_COMMON_END(ret, ret, blockresult);
-		PACK_RESOURCE(nWritten, (unsigned int)this, FD_FILE, FD_API_WRITE, 0, temp);
+		PACK_RESOURCE(nWritten, (unsigned long)this, FD_FILE, FD_API_WRITE, 0, temp);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -1132,7 +1132,7 @@ result File::Write(const Tizen::Base::String& buffer) {
 				  API_ID_result_File__Write_const_Tizen__Base__String__buffer_,
 				  "s", temp_buf);
 		PACK_COMMON_END(ret, ret, blockresult);
-		PACK_RESOURCE(nWritten, (unsigned int)this, FD_FILE, FD_API_WRITE, 0, temp_path);
+		PACK_RESOURCE(nWritten, (unsigned long)this, FD_FILE, FD_API_WRITE, 0, temp_path);
 		FLUSH_LOCAL_BUF();
 
 		 postBlockEnd();
@@ -1176,7 +1176,7 @@ File::~File(void) {
 				  API_ID_File___File_void_,
 				  "", 0);
 		PACK_COMMON_END(0, 0, blockresult);
-		PACK_RESOURCE(0, (unsigned int)this, FD_FILE, FD_API_CLOSE, 0, "");
+		PACK_RESOURCE(0, (unsigned long)this, FD_FILE, FD_API_CLOSE, 0, "");
 		FLUSH_LOCAL_BUF();
 
 		postBlockEnd();
